Change algorithm for constructing P, E
Individual threads are now in charge of their own coefficients. This
means that we no longer reduce to create P, rather we transform the 1s
and 0s representing whether a grid point is forcing or not into a tree
that represents the column of P that each thread corresponds to. This is
intended to make it so there is less branch divergence when constructing
A and P in the same kernel. Since the number of threads >= number of
rows >> number of columns, we supplement with a marker for whether a
thread has the same index as a forcing point within the larger grid.

There is however not much that can be done to remove branch divergence
in E, but it is still pretty fast.